API reference

This is the official API reference of PerfTest. Note that it can also be queried interactively from the Julia REPL using the help mode:

julia> using PerfTest
julia>?
help?> PerfTest

metrics type and basic constructors

PerfTest.metrics — Constant

Metrics configuration

regression, suppress_output, save_folder

PerfTest.regression — Constant

Regression testing configuration

PerfTest.suppress_output — Constant

Tested expressions output handling

PerfTest.save_folder — Constant

Where the regression tests are saved

Convenience type aliases and constructors

Index

PerfTest.retvalExpressionParser

PerfTest.metaGet

Documentation

PerfTest.ASTRule — Type

Used by the tree traverser to check for expressions that match "condition", if they do then "modifier" will be applied to the expression.

PerfTest.Context — Type

Saves important state information when going through the AST of an expression.

PerfTest.EnvironmentFlags — Type

Saves data needed during one specific execution of the test generation process.

Functions – additional to standard AbstractArray functionality

Index

PerfTest.configFallBack

PerfTest.getNumber

PerfTest.metaGet

PerfTest.retvalExpressionParser

PerfTest.rooflineMacroParse

Documentation

PerfTest.configFallBack — Method

A little automatism to jump to defaults if the configuration provided is absent Kind can be: :regression

PerfTest.getNumber — Method

WARNING, SUPPORTS ONLY 1 NUMBER PER LINE

PerfTest.metaGet — Method

Runs over an array of expressions trying to match the desired one. If not found returns "Nothing".

"sym" should follow the MacroTools nomenclature for the @capture macro

PerfTest.retvalExpressionParser — Method

Checks for the return symbol

PerfTest.rooflineMacroParse — Method

Parses roofline user request and sets up data for roofline computation.