API reference

This is the official API reference of PerfTest. Note that it can also be queried interactively from the Julia REPL using the help mode:

julia> using PerfTest
julia>?
help?> PerfTest

Types

Index

PerfTest.ASTRule

PerfTest.ASTWalkDepthRecord

PerfTest.Context

PerfTest.CustomMetric

PerfTest.DepthRecord

PerfTest.EnvironmentFlags

PerfTest.Methodology_Result

PerfTest.Metric_Constraint

PerfTest.Metric_Reference

PerfTest.Metric_Result

PerfTest.Perftest_Datafile_Root

PerfTest.Perftest_Result

PerfTest.Struct_Eff_Mem_Throughput

PerfTest.Struct_Metric_Config

PerfTest.Struct_Regression

PerfTest.Struct_Roofline_Config

PerfTest.Struct_Tolerance

Documentation

PerfTest.ASTRule — Type

Used by the AST walker to check for expressions that match condition, if they do then modifier will be applied to the expression.

This is the basic building block of the code transformer, a set of these rules compounds to all the needed manipulations to create the testing suite.

PerfTest.ASTWalkDepthRecord — Type

This structure is used to record a test set hierarchy during a AST walk. In any specific point of the walk the array will TODO

PerfTest.Context — Type

In order to perform with the test suite generation, the AST walk needs to keep a context register to integrate features that rely on the scope hierarchy.

PerfTest.CustomMetric — Type

Saves flags needed during the execution of the AST walk. It holds if:

The walk is on an expression that is a test target

The walk is on an expression that is inside a config macro

Several flags that affect the roofline methodology

PerfTest.DepthRecord — Type

This structure is used to record a test set frame during a AST walk. See ASTWalkDepthRecord for more info.

PerfTest.EnvironmentFlags — Type

Saves flags needed during the execution of the AST walk. It holds if:

The walk is on an expression that is a test target

The walk is on an expression that is inside a config macro

Several flags that affect the roofline methodology

PerfTest.Methodology_Result — Type

This struct is used in the test suite to save a methodology result, which in turn is constituted of a group of metric results and their references. Additionally, custom elements that are not subject to test are also saved, e.g. informational metrics, printing functions.

PerfTest.Metric_Constraint — Type

This struct is used in the test suite to save a metric test result and its associated data, it saves the reference used and the toreance intervals in absolute and percentual values, also it shows if the test succeded and some additional variables for data printing

PerfTest.Metric_Reference — Type

This struct is used in the test suite to save a metric reference, a reference is meant to be later compared with a result, its combination gives the Metric_Constraint struct. It holds:

A reference value.

low_is_bad registers if in this metric lower values are less desired than higher ones, or the opposite (e.g. time vs FLOP/s).

PerfTest.Metric_Result — Type

This struct is used in the test suite to save a metric measurement, therefore its saves the metric name, its units space and its value.

PerfTest.Perftest_Datafile_Root — Type

This struct is the root of the data recording file, it can save several performance test suite execution results.

PerfTest.Perftest_Result — Type

This struct saves a complete test suite result for one execution. It also saves the raw measurements obtained from the targets.

PerfTest.Struct_Eff_Mem_Throughput — Type

This struct holds the configuration of the basic effective memory throughput methodology.

enabled is used to enable or disable the methodology

tolerance defines the interval of ratios (eff.mem.through. / max. bandwidth) that make the test succeed.

PerfTest.Struct_Metric_Config — Type

This struct can hold the configuration of any metric.

enabled is used to enable or disable the methodology

regression_threshold, when comparing the measure with a reference, defines how far can the measurement be from the reference

PerfTest.Struct_Regression — Type

This struct holds the configuration of the basic metric regression methodology.

enabled is used to enable or disable the methodology save_failed will record historical measurements of failed tests if true general_regression_threshold sets the torelance interval for the test comparison

regression_calculation can be:

:latest The reference will be the latest saved result

:average The reference will be the average of all saved results

PerfTest.Struct_Roofline_Config — Type

This struct holds the configuration of the basic roofline methodology.

enabled is used to enable or disable the methodology

tolerance defines the interval of ratios (eff.mem.through. / max. bandwidth) that make the test succeed.

PerfTest.Struct_Tolerance — Type

Tolerance interval structure. Used to save intervals around a threshold during test comparisons.

Functions

Index

PerfTest._treeRun

PerfTest.autoflopExpressionParser

PerfTest.auxiliarMetricPrint

PerfTest.buildPrimitiveMetrics

PerfTest.by_index

PerfTest.checkAuxiliaryCustomMetrics

PerfTest.checkAuxiliaryMetric

PerfTest.checkCustomMetric

PerfTest.checkCustomMetrics

PerfTest.checkMedianTime

PerfTest.checkMinTime

PerfTest.configFallBack

PerfTest.customMetricExpressionParser

PerfTest.customMetricReferenceExpressionParser

PerfTest.extractMethodologyResultArray

PerfTest.extractNamesResultArray

PerfTest.flattenedInterpolation

PerfTest.fullParsingSuite

PerfTest.genTestName!

PerfTest.getMetricValue

PerfTest.getNumber

PerfTest.grepOutput

PerfTest.grepOutputXGetNumber

PerfTest.iteratorExpressionParser

PerfTest.loadFileAsExpr

PerfTest.metaGet

PerfTest.metaGetString

PerfTest.onCustomMetricDefinition

PerfTest.onMemoryThroughputDefinition

PerfTest.openDataFile

PerfTest.p_blue

PerfTest.p_green

PerfTest.p_red

PerfTest.p_yellow

PerfTest.perftestConfigEnter

PerfTest.perftestConfigExit

PerfTest.popQuoteBlocks

PerfTest.printDepth!

PerfTest.printIntervalLanding

PerfTest.printMethodology

PerfTest.printMetric

PerfTest.printedOutputExpressionParser

PerfTest.printfail

PerfTest.removeBlock

PerfTest.retvalExpressionParser

PerfTest.rooflineMacroParse

PerfTest.ruleSet

PerfTest.saveDataFile

PerfTest.saveExprAsFile

PerfTest.setupCPUPeakFlopBenchmark

PerfTest.setupMemoryBandwidthBenchmark

PerfTest.testsetUpdate!

PerfTest.treeRun

PerfTest.unblockAndConcat

PerfTest.@auxiliary_metric

PerfTest.@define_eff_memory_throughput

PerfTest.@define_metric

PerfTest.@lpad

PerfTest.@on_perftest_exec

PerfTest.@on_perftest_ignore

PerfTest.@perftest

PerfTest.@perftest_config

PerfTest.@roofline

Documentation

PerfTest._treeRun — Method

This method gets a input julia expression, and a context register and executes a transformation of the input that converts a recipe script (input) into a fully-fledged testing suite (return value).

Arguments

input_expr the recipe/source expression. (internally, a.k.a source code space)

context a register that will store information useful for the transformation over its run over the AST of the input

PerfTest.autoflopExpressionParser — Method

This is one of the parser functions that expand any formula block for metric definition. This function will parse the :autoflop symbol and substitute it with the flop count of the test target

PerfTest.auxiliarMetricPrint — Method

This function is used to dump metric information regading auxiliar metrics, which are not used in testing.

PerfTest.buildPrimitiveMetrics — Method

This function generates the code that make primitive metrics values available to all metodologies and custom metrics.

PerfTest.by_index — Method

This method expects a hierarchy tree (dict) in the form of nested dictionaries and a vector of dictionary keys idx. The function will recursively apply the keys to get to a final element.

It is usually put to work with the DepthRecord struct.

PerfTest.checkAuxiliaryCustomMetrics — Method

This method is used to generate the code that computes the value of every auxiliary custom metric enabled in th current context, where the code is generated.

PerfTest.checkAuxiliaryMetric — Method

This method is used to generate the code that computes the value of a given auxiliary custom metric in the context the code is generated.

PerfTest.checkCustomMetric — Method

This function is used to generate the code that evaluates if a custom metric result f a target is within a specified reference.

WARNING

Predefined symbols needed before this code is added to the generated space:

reference_value

metric_results

PerfTest.checkCustomMetrics — Method

This function is used to generate the code that evaluates if a custom metric result f a target is within a specified reference.

WARNING

Predefined symbols needed before this code is added to the generated space:

reference_value

metric_results

PerfTest.checkMedianTime — Method

This function is used to generate the code that evaluates if the median time of execution of a target is within a specified reference.

WARNING

Predefined symbols needed before this code is added to the generated space:

reference_value

metric_results

PerfTest.checkMinTime — Method

This function is used to generate the code that evaluates if the minimum time of execution of a target is within a specified reference.

WARNING

Predefined symbols needed before this code is added to the generated space:

reference_value

metric_results

PerfTest.configFallBack — Method

A little automatism to jump to defaults if the configuration provided is absent Kind can be: :regression

PerfTest.customMetricExpressionParser — Method

This is one of the parser functions that expand any formula block for metric definition. This function will parse all primitive metric symbols with the structure where the corresponding value of the metric is.

PerfTest.customMetricReferenceExpressionParser — Method

This is one of the parser functions that expand any formula block for metric definition. This function will parse all primitive metric symbols with the structure where the corresponding reference value for the metric is.

PerfTest.extractMethodologyResultArray — Method

This method will return a flattened array of all of the results for all the methodologies exercised in the provided dictionary.

Example:

"Test Set 1" -> "Test 1" -> Methodology A result -> Methodology B result "Test Set 2" -> "Test 1" -> Methodology A result Returns: M. A result (Test 1) M. B result (Test 1) M. A result (Test 2)

PerfTest.extractNamesResultArray — Method

This method will return a flattened array of the whole test result hierarchy.

Example

Example:

"Test Set 1" -> "Test 1" -> Methodology A result -> Methodology B result "Test Set 2" -> "Test 1" -> Methodology A result Returns: "Test Set 1 -> Test 1 -> Methodology A" "Test Set 1 -> Test 1 -> Methodology B" "Test Set 2 -> Test 1 -> Methodology A"

PerfTest.flattenedInterpolation — Method

This method interpolates the inside_expr into outside_expr anywhere it finds the token substitution_token, which is a symbol. The outside_expr has to be a block or a quote block. It has the particularity that it will remove block heads from the inside_expr and add the nested elements onto the location where the token it.

Example:

outside_expr = :(:A; 4)

inside_expr = :(begin 2;3 end)

substitution_token = :A

returns = :(2;3;4)

PerfTest.fullParsingSuite — Method

This function combines a collection of rules to turn a formula block into a functioning expression to calculate any metric defined by said formula

PerfTest.genTestName! — Method

Function that generates a test name if needed, it is used to name test targets to distinguish them if several go in the same testset.

PerfTest.getMetricValue — Method

Given a series of methodology results, the the raw values of all the metrics contained in the methodology results.

PerfTest.getNumber — Method

From a string (field), it will parse the first number it finds as a Float

PerfTest.grepOutput — Method

From a string, it will divide it by lines and retrieve the ones that match the regular expression provided.

PerfTest.grepOutputXGetNumber — Method

Given a string output, it will retrieve the first number in the first line that contains the string string.

PerfTest.iteratorExpressionParser — Method

This is one of the parser functions that expand any formula block for metric definition. This function will parse the :iterator symbol and substitute it with the current value of the innermost test set loop of the current test target execution

PerfTest.loadFileAsExpr — Method

Utility to get an expression from a Julia file stored at path

PerfTest.metaGet — Method

Runs over an array of expressions trying to match the desired one. If not found returns "Nothing".

"sym" should follow the MacroTools nomenclature for the @capture macro

PerfTest.metaGetString — Method

PerfTest.onCustomMetricDefinition — Method

This function is called to register a custom metric, it will parse the arguments of the definition macro and add the metric to the context to be later used in test targets on the same scope.

PerfTest.onMemoryThroughputDefinition — Method

This function is used to register a special custom metric, which is the effective memory throughput calculation, and is registered in the same way as any other but with a special flag that the EMT metholodogy will use to get and use the metric.

PerfTest.openDataFile — Method

This method is used to get historical data of a performance test suite from a save file located in path.

PerfTest.p_blue — Method

Prints the element in color blue

PerfTest.p_green — Method

Prints the element in color green

PerfTest.p_red — Method

Prints the element in color red

PerfTest.p_yellow — Method

Prints the element in color yellow

PerfTest.perftestConfigEnter — Method

Function to trigger the configuration mode on the context register

PerfTest.perftestConfigExit — Method

Function to deactivate the configuration mode on the context register

PerfTest.popQuoteBlocks — Method

Useful to correct operations limited by the tree walking Will remove quote blocks inside the main block without recursion and push their expressions into the main block

PerfTest.printDepth! — Method

This method is used to print the test names, with consideration on the hierarchy and adding indentation whenever necessary

PerfTest.printIntervalLanding — Function

This method is used to print a graphical representation on a test result and the admisible intervals it can take. The result will and the two bounds will be printed in order.

PerfTest.printMethodology — Method

This function is used to print the information relative to a methodology, relative to a a specific test execution result. This will usually print a series of metrics and might also print plots.

PerfTest.printMetric — Method

This method is used to dump into the output the information about a metric and the value obtained in a specific test.

PerfTest.printedOutputExpressionParser — Method

This is one of the parser functions that expand any formula block for metric definition. This function will parse the :printed_output symbol and substitute it with the standard output of the test target execution

PerfTest.printfail — Method

This method dumps into the output a test result in case of failure. The output will be formatted to make it easy to read.

PerfTest.removeBlock — Method

Pops expr which has a head that is :block or :quote and returns array of nested expressions which are the arguments of such head.

PerfTest.retvalExpressionParser — Method

This is one of the parser functions that expand any formula block for metric definition. This function will parse the appropiate symbol and substitute it by the return value of the test target execution.

PerfTest.rooflineMacroParse — Method

Parses roofline user request and sets up data for roofline computation.

PerfTest.ruleSet — Method

This method builds what is known as a rule set. Which is a function that will evaluate if an expression triggers a rule in a set and if that is the case apply the rule modifier. See the ASTRule documentation for more information.

WARNING: the rule set will apply the FIRST rule that matches with the expression, therefore other matches will be ignored

Arguments

context the context structure of the tree run, it will be ocassinally used by some rules on the set.

rules the collection of rules that will belong to the resulting set.

PerfTest.saveDataFile — Method

This method is used to save historical data of a performance test suite to a save file located in path.

PerfTest.saveExprAsFile — Function

Utility to save an expression (expr) to a Julia file stored at path

Requires a :toplevel symbol to be the head of the expression.

PerfTest.setupCPUPeakFlopBenchmark — Method

This method is used to generate the code responsible for sampling the maximum CPU FLOPS based on the avaiable threads in every resulting suite.

PerfTest.setupMemoryBandwidthBenchmark — Method

This method is used to generate the code responsible for sampling the maximum memory bandwith in every resulting suite.

PerfTest.testsetUpdate! — Method

Function used to register a new test set in the hierarchy record of the context, where name is the name of the test set.

PerfTest.treeRun — Method

This method implements the transformation that converts a recipe script into a fully-fledged testing suite. The function will return a Julia expression with the resulting performance testing suite. This can be then executed or saved in a file for later usage.

Arguments

path the path of the script to be transformed.

PerfTest.unblockAndConcat — Method

This function is useful to move expressions to the toplevel when they are enclosed inside a block

PerfTest.@auxiliary_metric — Macro

Defines a custom metric for informational purposes that will not be used for testing but will be printed as output.

PerfTest.@define_eff_memory_throughput — Macro

This macro is used to define the memory bandwidth of a target in order to execute the effective memory thorughput methodology.

Arguments

formula block : an expression that returns a single value, which would be the metric value. The formula can have any julia expression inside and additionally some special symbols are supported. The formula may be evaluated several times, so its applied to every target in every test set or just once, if the formula is defined inside a test set, which makes it only applicable to it.

Special symbols:

:median_time : will be substituted by the median time the target took to execute in the benchmark.

:minimum_time: will be substituted by the minimum time the target took to execute in the benchmark.

:ret_value : will be substituted by the return value of the target.

:autoflop: will be substituted by the FLOP count the target.

:printed_output : will be substituted by the standard output stream of the target.

:iterator : will be substituted by the current iterator value in a loop test set.

Example:

The following definition assumes that each execution of the target expression involves transacting 1000 bytes. Therefore the bandwith is 1000 / execution time.

@define_eff_memory_throughput begin
      1000 / :median_time
end

PerfTest.@define_metric — Macro

This macro is used to define a new custom metric.

Arguments

name : the name of the metric for identification purposes.

units : the unit space that the metric values will be in.

formula block : an expression that returns a single value, which would be the metric value. The formula can have any julia expression inside and additionally some special symbols are supported. The formula may be evaluated several times, so its applied to every target in every test set or just once, if the formula is defined inside a test set, which makes it only applicable to it.

Special symbols:

:median_time : will be substituted by the median time the target took to execute in the benchmark.

:minimum_time: will be substituted by the minimum time the target took to execute in the benchmark.

:ret_value : will be substituted by the return value of the target.

:autoflop: will be substituted by the FLOP count the target.

:printed_output : will be substituted by the standard output stream of the target.

:iterator : will be substituted by the current iterator value in a loop test set.

PerfTest.@lpad — Macro

Macro that adds a space at the beggining of a string

PerfTest.@on_perftest_exec — Macro

The expression given to this macro will only be executed in the generated suite, and will be deleted if the source code is executed as is.

PerfTest.@on_perftest_ignore — Macro

The expression given to this macro will only be executed in the source code, and will be deleted in the generated performance test suite.

PerfTest.@perftest — Macro

This macro is used to signal that the wrapped expression is a performance test target, and therefore its performance will be sampled and then evaluated following the current suite configuration.

If the macro is evaluated it does not modify the target at all. The effects of the macro only show when the script is transformed into a performance testing suite.

This macro is sensitive to context since other adjacent macros can change how the target will be evaluated.

Arguments

The target expression

Example

@perftest 2 + 3

PerfTest.@perftest_config — Macro

Perftest_config macro, used to set customised configuration on the suite generated by the source script

Configuration inside this macro must follow the syntax below:

@perftest_config
    key = value
    key.subkey = value
end

Where key can be any configuration parameter, in some cases parameters will consist on a set of subparameters denoted by the "." to refer to them.

PerfTest.@roofline — Macro

This macro enables roofline modelling, if put just before a target declaration (@perftest) it will proceed to evaluate it using a roofline model.

Mandatory arguments

formula block: the macro has to wrap a block that holds a formula to obtain the operational intensity of target algorithms.

Optional arguments

cpu_peak : a manual input value for the maximum attainable FLOPS, this will override the empirical runtime benchmark

membw_peak : a manual input value for the maximum memory bandwith, this will override the empirical runtime benchmark

target_opint : a desired operational intensity for the target, this will turn operational intensity into a test metric

actual_flops: another formula that defines the actual performance of the test.

target_ratio : the acceptable ratio between the actual performance and the projected performance from the roofline, this will turn actual performance into a test metric.

Special symbols:

:median_time : will be substituted by the median time the target took to execute in the benchmark.

:minimum_time: will be substituted by the minimum time the target took to execute in the benchmark.

:ret_value : will be substituted by the return value of the target.

:autoflop: will be substituted by the FLOP count the target.

:printed_output : will be substituted by the standard output stream of the target.

:iterator : will be substituted by the current iterator value in a loop test set.

Any formula block specified in this macro supports these symbols.

Example

@roofline actual_flops=:autoflop target_ratio=0.05 begin
    mem = ((:iterator + 1) * :iterator)
    :autoflop / mem
end

The code block defines operational intensity, whilst the other arguments define how to measure and compare the actual performance with the roofline performance. If the actual to projected performance ratio goes below the target, the test fails.